requirements.dev.txt
inmanta-dev-dependencies[module]==2.58.0
